requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
anthropic==0.49.0
python-dotenv==1.0.1
